Added methods for renderers to discover needed properties of plugin (support for setTime, min/max image inputs, etc).

+++ vvopensource/VVFFGL/FrameworkSrc/FFGLRenderer.h
@@ -13,7 +13,7 @@
 @interface FFGLRenderer : NSObject {
 @private
     FFGLPlugin          *_plugin;
-    NSUInteger          _instance;
+    uint32_t            _instance;
     CGLContextObj       _pluginContext;
     NSRect              _bounds;
     NSString            *_pixelFormat;
